refactor(widget-plugin-rollup-build-ts): move context to new file
